refactor(ios): split SpotifyAppRemoteCoordinator into focused modules
Extract error types, SPT error mapping, and JS bridge mappers so the coordinator stays connection/API-only and stays under the 1k-line threshold as App Remote grows.
